# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The MelonWallet microservice responsible for dealing with payment and crypto transaction information.",
        "title": "Payment Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/transaction": {
            "get": {
                "description": "List all transactiones of an account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List all transactiones of an account",
                "operationId": "list-transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github.com_Melon-Network-Inc_payment-service_pkg_transaction.Transaction"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            },
            "put": {
                "description": "Update a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Update a transaction",
                "operationId": "update-transaction",
                "parameters": [
                    {
                        "description": "Transaction Data",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_Melon-Network-Inc_payment-service_pkg_transaction.UpdateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_Melon-Network-Inc_payment-service_pkg_transaction.Transaction"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    }
                }
            },
            "post": {
                "description": "Add a transaction to account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "addresses"
                ],
                "summary": "Add a transaction to account",
                "operationId": "add-address",
                "parameters": [
                    {
                        "description": "Transaction Data",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_Melon-Network-Inc_payment-service_pkg_transaction.AddTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github.com_Melon-Network-Inc_payment-service_pkg_transaction.Transaction"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Delete a transaction",
                "operationId": "delete-transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_Melon-Network-Inc_payment-service_pkg_transaction.Transaction"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/transaction/{id}": {
            "get": {
                "description": "Get an transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get an transaction",
                "operationId": "get-transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_Melon-Network-Inc_payment-service_pkg_transaction.Transaction"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "github.com_Melon-Network-Inc_payment-service_pkg_transaction.AddTransactionRequest": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "name",
                "receiver_pk",
                "sender_pk"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "description": "currency code",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "integer"
                },
                "receiver_pk": {
                    "description": "ETH or BTC address",
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_pk": {
                    "description": "ETH or BTC address",
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github.com_Melon-Network-Inc_payment-service_pkg_transaction.Transaction": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "name",
                "receiver_pk",
                "sender_pk"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "description": "currency code",
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "description": "message should be less than 200 characters",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "integer"
                },
                "receiver_pk": {
                    "description": "ETH or BTC address",
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_pk": {
                    "description": "ETH or BTC address",
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github.com_Melon-Network-Inc_payment-service_pkg_transaction.UpdateTransactionRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "pkg_transaction.AddTransactionRequest": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "name",
                "receiver_pk",
                "sender_pk"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "description": "currency code",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "integer"
                },
                "receiver_pk": {
                    "description": "ETH or BTC address",
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_pk": {
                    "description": "ETH or BTC address",
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "pkg_transaction.Transaction": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "name",
                "receiver_pk",
                "sender_pk"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "description": "currency code",
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "description": "message should be less than 200 characters",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "integer"
                },
                "receiver_pk": {
                    "description": "ETH or BTC address",
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_pk": {
                    "description": "ETH or BTC address",
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "pkg_transaction.UpdateTransactionRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}